[WEB-4189] - renamed text campaign to outreach
